Create First go at code for covid-19
So I found these SIR deferential equations which are used to find the spread of a disease through a population. I than use a def and a odient function solve them and than graph. the only thing I can think of to add is a morality graph and or any other stuff anyone can think of.